Implemented the Warp skill.
When the player presses the Triangle button, Warp is activated and a
crosshair appears on the screen. Upon a second Triangle button press,
the player warps/teleports/etc. to the location of the crosshair.

Activation of the skill costs 50% of Energy and warping will deplete the
rest of the Energy bar. If the player does not press the Triangle button
a second time after activating in a time interval of 2.5 seconds, Warp
deactivates without warping the player, and without depleting the rest
of the Energy bar (only the 50% activation Energy cost).

The distance of the crosshair relative to the player's current position
(i.e. the warping distance) increases with the number of skill points
invested in Warp as follows:
Lv1-Lv5: 100/110/125/150/180 pixels.