1. Implemented the Force Field skill.
When the player presses the Cross button, Force Field is activated and
the ship will be protected from impacts while regenerating Energy.

Activation of the skill costs 50/55/60/65/70 (Lv1-Lv5) and lasts either
1/2/3/4/5 second (Lv1-Lv5) or for 1/2/3/4/5 (Lv1-Lv5) impacts. Although
generating the shield costs a fair amount of the Drive Core's Energy, it
is able to convert the kinetic energy of impacts with space debris back
into 10% of maximum Energy per impact.